Convert rdfxml into json for elasticsearch
Created a custom FimsModel for PPO-Fims. This correctly loops over all
relations and flattens out the graph into a json object that can be
loaded into es.

Performance is exponentially better as we removed the use of apache poi.

updated the ProjectController refreshCache service to add a "types"
field mapping in the generate es mapping.